rename scope to be in ibis.expr.scope
