+++ ibis/pandas/tests/test_core.py
@@ -9,7 +9,7 @@ import ibis
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
-from ibis.common.scope import Scope
+from ibis.expr.scope import Scope
 from ibis.pandas.client import PandasClient
 from ibis.pandas.core import is_computable_input
 from ibis.pandas.dispatch import execute_node, post_execute, pre_execute
